feat: removed all claude code
